cleaned some stuff
