+++ src/main/kotlin/edu/ics499/team1/app/AppApplication.kt
@@ -2,25 +2,10 @@ package edu.ics499.team1.app
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
-import org.springframework.test.context.web.WebAppConfiguration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
 @SpringBootApplication
 class Application
 
 fun main(args: Array<String>) {
     runApplication<Application>(*args)
-}
-
-
-@WebAppConfiguration
-class WebConfig : WebMvcConfigurer {
-    override fun addCorsMappings(registry: CorsRegistry) {
-        println("addCorsMappings method is being called!")
-        registry.addMapping("/**")
-            .allowedOrigins("http://localhost:5173")
-            .allowedMethods("GET", "POST", "PATCH", "DELETE")
-            .allowedHeaders("*")
-    }
 }
